Feature/update product service

+++ services/order/src/main/java/dev/berke/app/product/PurchaseResponse.java
@@ -3,11 +3,9 @@ package dev.berke.app.product;
 import java.math.BigDecimal;
 
 public record PurchaseResponse(
-
         Integer productId,
         String name,
         String description,
         BigDecimal price,
-        double quantity
-) {
+        Integer quantity) {
 }
