+++ services/product/src/main/java/dev/berke/app/product/Product.java
@@ -29,11 +29,10 @@ public class Product {
     private Integer id;
     private String name;
     private String description;
-    private double availableQuantity;
+    private Integer availableQuantity;
     private BigDecimal price;
 
     @ManyToOne
     @JoinColumn(name = "category_id")
     private Category category;
 }
-
