+++ services/product/src/main/java/dev/berke/app/product/ProductPurchaseResponse.java
@@ -3,11 +3,10 @@ package dev.berke.app.product;
 import java.math.BigDecimal;
 
 public record ProductPurchaseResponse(
-
         Integer productId,
         String name,
         String description,
         BigDecimal price,
-        double quantity
+        Integer quantity
 ) {
 }
