+++ services/product/src/main/java/dev/berke/app/product/ProductResponse.java
@@ -3,14 +3,12 @@ package dev.berke.app.product;
 import java.math.BigDecimal;
 
 public record ProductResponse(
-
-                Integer id,
-                String name,
-                String description,
-                double availableQuantity,
-                BigDecimal price,
-                Integer categoryId,
-                String categoryName,
-                String categoryDescription
-) {
+        Integer id,
+        String name,
+        String description,
+        Integer availableQuantity,
+        BigDecimal price,
+        Integer categoryId,
+        String categoryName,
+        String categoryDescription) {
 }
